perf(partners): phase-time BOTH partner create routes (#1370)
Follow-up within the same PR: the first pass instrumented only the legacy
`POST /partners/products`, which is what the assistant's `create_product` tool
posts to and what my prod repro happened to hit. But the partner UI creates
through the store-scoped twin `POST /partners/stores/:id/products`
(partner-ui/src/hooks/api/products.tsx:428) — that route's own comment says so:
"This is the route the partner-ui actually calls".

Both run the same post-write sequence (createProductsWorkflow ->
ensureInventoryLevelsForVariants -> requestVariantPriceFanout), so instrumenting
one and reasoning about the other is exactly how the partner-facing path stays
unmeasured.

`phaseTimer` now lives in the shared partner helpers instead of being copied per
route, and takes the log label as an argument.

The store-scoped route additionally marks `proposalGate` and `proposalRecord` —
the artisan-proposal path is unique to it and does a link write plus an event
emit, so it needs to be visible rather than folded into a neighbour.

variants/batch keeps its own local copy for now: it is the route under active
measurement and changing it would add a variable.

Verified: tsc clean on both routes, check:prod-build passed.
